chore(mps-sync-plugin): add 'plugin' to the package FQN
