+++ mps-sync-plugin/src/main/kotlin/org/modelix/mps/sync/plugin/ModelSyncService.kt
@@ -1,4 +1,20 @@
-package org.modelix.mps.sync
+/*
+ * Copyright (c) 2024.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.modelix.mps.sync.plugin
 /*
  * Copyright (c) 2023.
  *
@@ -30,7 +46,8 @@ import org.modelix.model.api.runSynchronized
 import org.modelix.model.client2.ModelClientV2
 import org.modelix.model.lazy.BranchReference
 import org.modelix.model.lazy.RepositoryId
-import org.modelix.mps.sync.action.ModelixActionGroup
+import org.modelix.mps.sync.SyncServiceImpl
+import org.modelix.mps.sync.plugin.action.ModelixActionGroup
 import java.net.ConnectException
 import java.net.URL
