+++ mps-sync-plugin/src/main/kotlin/org/modelix/mps/sync/plugin/action/ModelSyncAction.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023.
+ * Copyright (c) 2023-2024.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,31 +14,31 @@
  * limitations under the License.
  */
 
-package org.modelix.mps.sync.action
+package org.modelix.mps.sync.plugin.action
 
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.DataKey
 import com.intellij.openapi.diagnostic.logger
-import jetbrains.mps.project.AbstractModule
-import org.jetbrains.mps.openapi.module.SModule
+import jetbrains.mps.extapi.model.SModelBase
+import org.jetbrains.mps.openapi.model.SModel
 import org.modelix.kotlin.utils.UnstableModelixFeature
 import org.modelix.mps.sync.bindings.BindingsRegistry
 import org.modelix.mps.sync.modelix.ReplicatedModelRegistry
 import org.modelix.mps.sync.transformation.cache.MpsToModelixMap
-import org.modelix.mps.sync.transformation.mpsToModelix.initial.ModuleSynchronizer
+import org.modelix.mps.sync.transformation.mpsToModelix.initial.ModelSynchronizer
 import org.modelix.mps.sync.util.SyncQueue
 
 @UnstableModelixFeature(reason = "The new modelix MPS plugin is under construction", intendedFinalization = "2024.1")
-class ModuleSyncAction : AnAction {
+class ModelSyncAction : AnAction {
 
     companion object {
-        val CONTEXT_MODULE = DataKey.create<SModule>("MPS_Context_SModule")
+        val CONTEXT_MODEL = DataKey.create<SModel>("MPS_Context_SModel")
 
-        fun create() = ModuleSyncAction("Synchronize module to server")
+        fun create() = ModelSyncAction("Synchronize model to server")
     }
 
-    private val logger = logger<ModuleSyncAction>()
+    private val logger = logger<ModelSyncAction>()
 
     constructor() : super()
 
@@ -46,11 +46,12 @@ class ModuleSyncAction : AnAction {
 
     override fun actionPerformed(event: AnActionEvent) {
         try {
-            val module = event.getData(CONTEXT_MODULE)!! as AbstractModule
+            val model = event.getData(CONTEXT_MODEL)!! as SModelBase
             val branch = ReplicatedModelRegistry.model!!.getBranch()
-            ModuleSynchronizer(branch, MpsToModelixMap, BindingsRegistry, SyncQueue).addModule(module)
+            // TODO fixme: warn the user if the model imports another model that is not on the model server yet
+            ModelSynchronizer(branch, MpsToModelixMap, BindingsRegistry, SyncQueue).addModelAndActivate(model)
         } catch (ex: Exception) {
-            logger.error("Module sync error occurred", ex)
+            logger.error("Model sync error occurred", ex)
         }
     }
 }
