+++ mps-sync-plugin/src/main/kotlin/org/modelix/mps/sync/plugin/action/ModuleSyncAction.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023.
+ * Copyright (c) 2023-2024.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,31 +14,31 @@
  * limitations under the License.
  */
 
-package org.modelix.mps.sync.action
+package org.modelix.mps.sync.plugin.action
 
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.DataKey
 import com.intellij.openapi.diagnostic.logger
-import jetbrains.mps.extapi.model.SModelBase
-import org.jetbrains.mps.openapi.model.SModel
+import jetbrains.mps.project.AbstractModule
+import org.jetbrains.mps.openapi.module.SModule
 import org.modelix.kotlin.utils.UnstableModelixFeature
 import org.modelix.mps.sync.bindings.BindingsRegistry
 import org.modelix.mps.sync.modelix.ReplicatedModelRegistry
 import org.modelix.mps.sync.transformation.cache.MpsToModelixMap
-import org.modelix.mps.sync.transformation.mpsToModelix.initial.ModelSynchronizer
+import org.modelix.mps.sync.transformation.mpsToModelix.initial.ModuleSynchronizer
 import org.modelix.mps.sync.util.SyncQueue
 
 @UnstableModelixFeature(reason = "The new modelix MPS plugin is under construction", intendedFinalization = "2024.1")
-class ModelSyncAction : AnAction {
+class ModuleSyncAction : AnAction {
 
     companion object {
-        val CONTEXT_MODEL = DataKey.create<SModel>("MPS_Context_SModel")
+        val CONTEXT_MODULE = DataKey.create<SModule>("MPS_Context_SModule")
 
-        fun create() = ModelSyncAction("Synchronize model to server")
+        fun create() = ModuleSyncAction("Synchronize module to server")
     }
 
-    private val logger = logger<ModelSyncAction>()
+    private val logger = logger<ModuleSyncAction>()
 
     constructor() : super()
 
@@ -46,12 +46,11 @@ class ModelSyncAction : AnAction {
 
     override fun actionPerformed(event: AnActionEvent) {
         try {
-            val model = event.getData(CONTEXT_MODEL)!! as SModelBase
+            val module = event.getData(CONTEXT_MODULE)!! as AbstractModule
             val branch = ReplicatedModelRegistry.model!!.getBranch()
-            // TODO fixme: warn the user if the model imports another model that is not on the model server yet
-            ModelSynchronizer(branch, MpsToModelixMap, BindingsRegistry, SyncQueue).addModelAndActivate(model)
+            ModuleSynchronizer(branch, MpsToModelixMap, BindingsRegistry, SyncQueue).addModule(module)
         } catch (ex: Exception) {
-            logger.error("Model sync error occurred", ex)
+            logger.error("Module sync error occurred", ex)
         }
     }
 }
