+++ mps-sync-plugin/src/main/kotlin/org/modelix/mps/sync/plugin/gui/BindingsComboBoxRefresher.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.modelix.mps.sync.gui
+package org.modelix.mps.sync.plugin.gui
 
 import org.modelix.kotlin.utils.UnstableModelixFeature
 import org.modelix.mps.sync.IBinding
