+++ mps-sync-plugin/src/main/kotlin/org/modelix/mps/sync/plugin/indicator/CloudIndicator.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023.
+ * Copyright (c) 2023-2024.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.modelix.mps.sync.indicator
+package org.modelix.mps.sync.plugin.indicator
 
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.wm.CustomStatusBarWidget
@@ -26,7 +26,7 @@ import com.intellij.ui.UIBundle
 import com.intellij.ui.components.JBLabel
 import org.jetbrains.annotations.Nls
 import org.modelix.kotlin.utils.UnstableModelixFeature
-import org.modelix.mps.sync.icons.CloudIcons
+import org.modelix.mps.sync.plugin.icons.CloudIcons
 import java.awt.event.MouseEvent
 import javax.swing.JComponent
